Stop the search for photographs from outliving the folder that asked
Every folder open, every F5 and every theme change started an uncancellable
walk of the whole subtree. A Movies root of four hundred folders is the worst
case: each poster is claimed as artwork, so no photograph is ever found, and
the walk descends into every season and extras folder to the depth cap over
SMB only to answer "no" — and navigating away starts another rather than
stopping the one already running.

Three changes together. A folder whose own scan already produced an image
needs no walk at all, so BrowseView answers from what it has and inserts the
tile through the same path the late answer uses. hasPhotos now takes the
cancellation supplier collect has always taken instead of hard-coding one that
never fires. BrowseView implements onHidden and passes a supplier that stops
the walk when the page is gone or a newer load has begun; loadGeneration is
volatile because the walker thread is what reads it.

Alongside that, four smaller things the same review found. PhotoCache.clear
releases twenty-five megabytes on the one exit path that matters and had no
test, only inspection. ExifOrientation's bounds check summed in int, so an
offset near the top of the range wrapped negative and slipped past the guard
into the catch the comment claims it never reaches. The Slideshow settings row
built a value label and never filled it, leaving a hand-edited interval such
as 30 invisible: neither toggle selected and nothing on screen saying why. And
ArtworkCache dropped evicted images without cancelling them, so opening a
folder of five hundred photographs abandoned two hundred and sixty decodes
still reading the share for thumbnails nothing holds any more.

+++ src/main/java/mediacenter/media/ExifOrientation.java
@@ -70,7 +70,9 @@ public final class ExifOrientation {
 
         int ifdOffset = buffer.getInt(tiff + 4);
         // Checked rather than caught: a corrupt offset can be negative or enormous.
-        if (ifdOffset < 8 || tiff + ifdOffset + 2 > header.length) {
+        // The sum is taken in long because an offset near Integer.MAX_VALUE would
+        // otherwise wrap negative and slip past the very check meant to stop it.
+        if (ifdOffset < 8 || (long) tiff + ifdOffset + 2 > header.length) {
             return 0;
         }
         int ifd = tiff + ifdOffset;
